Add the Spring GraphQL dependencies
This commit adds the main and test Spring GraphQL dependencies to the
project. A GraphQL application needs a schema defined or it will fail to
start. This commit also adds a `/graphql/schema.graphqls` file with a
`greeting` query.

Once the application started, we can access the GraphiQL web console at
http://localhost:8080/graphiql and explore our schema.

Of course, the newly defined query will not work if we use it since no
implementation has been provided.