Fetch contributors from the GitHub API
This commit adds a new data source that's using the GitHub API to fetch
the list of contributors for a given Spring project.

There can be many calls to the GitHub API so we need to configure a
personal access token with our application or we may risk hitting the
default rate limiting.